Tested on last year's 15 inch robot. Still issues with shooter up to speed code but currently being attributed to value tuning.

+++ SingleFlyWheelShooter/include/SingleFlyWheelShooter.h
@@ -15,7 +15,7 @@ struct SingleFlyWheelShooter{
 	IME ime;
 	int on;
 	long lastOffTime;
-	int processVariable;
+	double processVariable;
 
 }typedef SingleFlyWheelShooter;
